Add support for stunnel 5.57 with wolfSSL.
I created a new directory hierarchy under stunnel/. 5.40 holds the old stuff,
which was a full copy of stunnel 5.40 modified to work with wolfssl. 5.57 holds
the new stuff, primarily a patch file to apply to a copy of stunnel 5.57 that
modifies the source code to work with wolfSSL. This is more lightweight than
keeping a full, modified copy of stunnel 5.57 around in this repo, and clearly
allows you to see the differences between the stunnel 5.57 base and the wolfSSL
port.