Make lighttpd-1.4.51 work with latest wolfSSL master.
Our integration with this version is currently broken because we now define the
macro SSL_CONF_FLAG_CMDLINE whereas we didn't in 4.7.0. That causes new code in
mod_openssl.c to be included. In that code, lighty calls SSL_CONF_CTX_set_flags
with flags we don't support. The solution for now is to ensure that this code
isn't compiled in the first place, as was the case back in 4.7.0.