Fix OpenSSH 8.5p1 patch to call EVP_Cipher "final" for AES-GCM.
Calling EVP_Cipher will NULL input is effectively like an EVP_CipherFinal call.
With OpenSSL's AES-GCM implementation, this will compute the tag on encrypt and
check the tag on decrypt. This isn't quite the same behavior as wolfSSL, but as
of wolfSSL PR 5170, the call is needed nonetheless. Previously, we had ifdef'd
the call out, as it would throw an error prior to this PR.

Additionally, I tweaked integrity.sh so that it passes with wolfSSL. That test
expects MAC errors to be reported as such, but with wolfSSL they're reported as
generic "libcrypto" errors. This needs to be accounted for, or the test fails.